+++ src/textual_textarea/cancellable_input.py
@@ -1,8 +1,0 @@
-from textual.widgets import Input
-
-
-class CancellableInput(Input):
-    BINDINGS = [("escape", "cancel", "Cancel")]
-
-    def action_cancel(self) -> None:
-        self.remove()